database-helper/pyproject.toml
[project]
name = "database-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pika>=1.3.2",
    "psycopg2-binary>=2.9.10",
]
